+++ src/hooks/register-server.ts
@@ -14,7 +14,6 @@ export default class RegisterServer {
           susChance: 0.01,
           randomSus: false,
           russianRoulette: false,
-          reactionRoles: false,
         },
         gulag: [],
         reactionRoles: [],