+++ changelog.md
@@ -3,6 +3,7 @@
 ## 1.2.2
  - fix types for command arguments
  - add gulag leaderboard
+ - remove reactionRoles setting
 
 ## 1.2.1
 - optimize gulag command